getting acceptance tests working again
